Move math functions into separate file
